Manifest: Add shared=ipc  (opencpn/opencpn#3601)
According to https://github.com/OpenCPN/OpenCPN/issues/3601, KDE
on X11 reveals a missing share=ipc in the manifest.

Documentation says that while ipc is not strictly required for X11,
omitting it comes with a performance penalty. Let's thus enable it
for now, a large part of  the user base is still on X11.